Fix setOrder which should take an unsigned long

+++ extensions/sld/include/powsybl/iidm/extensions/sld/ConnectablePosition.hpp
@@ -44,7 +44,7 @@ public:
 
         Feeder& setName(const std::string& name);
 
-        Feeder& setOrder(int order);
+        Feeder& setOrder(unsigned long order);
 
     private:
         std::string m_name;
